flashdev: fix wrong offset assignment
When the region feature is active, flashdev assumes that
a relative address is provided. It computes the abs. address and
carries out the read / write. However, in this case it must not
assign the abs. address to iop->offset. The relative address is
required here to allow newlib to correctly read/write the next
junk of data.

Updates #4981