flashdev: Refactor RTEMS_FLASHDEV_IOCTL_GET_WRITE_BLOCK_SIZE
The major change in this patch is that it refactors
RTEMS_FLASHDEV_IOCTL_GET_WRITE_BLOCK_SIZE to become
RTEMS_FLASHDEV_IOCTL_GET_MIN_WRITE_BLOCK_SIZE.

Apart from that this patch fixes comments and aligns function names.

There is no change in the behaviour of the code introduced by this
patch.

Updates #4981